wheather_app/requirements.txt
Django==5.0.7
PyYAML==6.0.1
asgiref==3.8.1
black==24.4.2
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
djangorestframework==3.15.2
filelock==3.15.4
flake8==7.1.0
flake8-broken-line==1.0.0
flake8-isort==6.1.1
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
gunicorn==22.0.0
geopandas==1.0.1
identify==2.6.0
idna==3.7
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
networkx==3.3
nodeenv==1.9.1
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pep8-naming==0.14.1
pip==24.1.2
platformdirs==4.2.2
pre-commit==3.7.1
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pyflakes==3.2.0
pyogrio==0.9.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
setuptools==65.5.0
shapely==2.0.5
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
